new install verify
