--- pgdrive/tests/install_test/test_get_image_headless.py
@@ -1,5 +1,0 @@
-from .test_get_image import capture_image
-
-if __name__ == "__main__":
-    capture_image(True)
-    print("Headless mode Offscreen render launched successfully! Images are saved, Please check them.")
