+++ pgdrive/tests/install_test/test_headless_install.py
@@ -1,0 +1,7 @@
+from .get_image import test_install
+from panda3d.core import loadPrcFileData
+loadPrcFileData("", "notify-level-task fatal")
+
+if __name__ == "__main__":
+    test_install(True)
+
